+++ src/Bridge/BridgeClient.ts
@@ -1,20 +1,10 @@
-import * as AsyncLock from 'async-lock';
-import ISocket from '@signageos/front-display/es6/Socket/ISocket';
-import Orientation from '@signageos/front-display/es6/NativeDevice/Orientation';
-import BridgeVideoClient from './BridgeVideoClient';
+export class BridgeRequestFailedError {}
 
 export default class BridgeClient {
 
-	public video: BridgeVideoClient;
-
 	constructor(
 		private serverUri: string,
-		public readonly socketClient: ISocket,
 	) {}
-
-	public initialize(window: Window, getOrientation: () => Orientation, lock: AsyncLock) {
-		this.video = new BridgeVideoClient(window, getOrientation, lock, this.socketClient);
-	}
 
 	public async invoke<TMessage extends { type: string }, TResult>(message: TMessage): Promise<TResult> {
 		const response = await fetch(
@@ -29,9 +19,56 @@ export default class BridgeClient {
 		);
 
 		if (!response.ok) {
-			throw new Error('bridge invoke failed with status code ' + response.status);
+			throw new BridgeRequestFailedError();
 		}
 
 		return await response.json();
+	}
+
+	public async uploadOverlay(
+		fileBlob: Blob,
+		id: string,
+		width: number,
+		height: number,
+		x: number,
+		y: number,
+		animation?: {
+			duration: number;
+			keyframes: {
+				percentage: number;
+				rectangle: {
+					x: number;
+					y: number;
+				};
+			}[];
+		},
+	) {
+		const params: { [key: string]: any } = { id, width, height, x, y };
+		if (animation && animation.keyframes.length > 1) {
+			params.animDuration = animation.duration;
+			params.animKFCount = animation.keyframes.length;
+			for (let i = 0; i < animation.keyframes.length; i++) {
+				params['animKF' + i + '_percent'] = animation.keyframes[i].percentage;
+				params['animKF' + i + '_x'] = animation.keyframes[i].rectangle.x;
+				params['animKF' + i + '_y'] = animation.keyframes[i].rectangle.y;
+			}
+		}
+
+		const paramsString = Object.keys(params)
+			.filter((key: string) => typeof params[key] !== 'undefined')
+			.map((key: string) => `${key}=${params[key]}`)
+			.join('&');
+
+		const response = await fetch(
+			this.serverUri + '/overlay?' + paramsString,
+			{
+				method: 'POST',
+				body: fileBlob,
+			},
+		);
+
+		if (!response.ok) {
+			throw new Error('Failed to upload overlay image with status code ' + response.status);
+		}
 	}
 }