+++ src/Bridge/bridgeOverlayMessages.ts
@@ -1,0 +1,10 @@
+export const Hide = 'Overlay.Hide';
+export interface Hide {
+	type: typeof Hide;
+	id: string;
+}
+
+export const HideAll = 'Overlay.HideAll';
+export interface HideAll {
+	type: typeof HideAll;
+}