+++ src/FileSystem/FileSystem.ts
@@ -104,7 +104,7 @@ export default class FileSystem implements IFileSystem {
 		return contents.toString();
 	}
 
-	public async saveToFile(filePath: IFilePath, contents: string) {
+	public async saveToFile(filePath: IFilePath, contents: string | Buffer) {
 		const fullPath = this.getAbsolutePath(filePath);
 		return await fs.writeFile(fullPath, contents);
 	}