+++ test/unit/Bridge/handleMessageSpec.ts
@@ -39,7 +39,7 @@ describe('Bridge.handleMessage', function () {
 		const nativeDriver = {
 			getDeviceUid: sinon.stub().resolves('deviceUid1'),
 		};
-		const result = await handleMessage({} as any, nativeDriver as any, { type: GetDeviceUid } as GetDeviceUid);
+		const result = await handleMessage({} as any, nativeDriver as any, {} as any, { type: GetDeviceUid } as GetDeviceUid);
 		result.should.deepEqual({ deviceUid: 'deviceUid1' });
 	});
 
@@ -51,7 +51,7 @@ describe('Bridge.handleMessage', function () {
 				getFilePath('directory/file3'),
 			]),
 		};
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: ListFiles,
 			directoryPath: getFilePath('directory'),
 		});
@@ -68,7 +68,7 @@ describe('Bridge.handleMessage', function () {
 		const fileSystem = {
 			exists: sinon.stub().resolves(true),
 		};
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: FileExists,
 			filePath: getFilePath('file1'),
 		});
@@ -78,7 +78,7 @@ describe('Bridge.handleMessage', function () {
 	it('should process DownloadFile message', async function () {
 		const downloadFile = sinon.stub().resolves();
 		const fileSystem = { downloadFile };
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: DownloadFile,
 			filePath: getFilePath('file1'),
 			sourceUri: 'uri1',
@@ -90,7 +90,7 @@ describe('Bridge.handleMessage', function () {
 	it('should process DeleteFile message', async function () {
 		const deleteFile = sinon.stub().resolves();
 		const fileSystem = { deleteFile };
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: DeleteFile,
 			filePath: getFilePath('file1'),
 			recursive: false,
@@ -102,7 +102,7 @@ describe('Bridge.handleMessage', function () {
 	it('should process MoveFile message', async function () {
 		const moveFile = sinon.stub().resolves();
 		const fileSystem = { moveFile };
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: MoveFile,
 			sourceFilePath: getFilePath('source'),
 			destinationFilePath: getFilePath('destination'),
@@ -115,7 +115,7 @@ describe('Bridge.handleMessage', function () {
 		const fileSystem = {
 			getFileChecksum: async () => 'result_checksum',
 		};
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: GetFileChecksum,
 			filePath: getFilePath('file'),
 			hashType: 'md5' as any,
@@ -126,7 +126,7 @@ describe('Bridge.handleMessage', function () {
 	it('should process ExtractFile message', async function () {
 		const extractFile = sinon.stub().resolves();
 		const fileSystem = { extractFile };
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: ExtractFile,
 			archiveFilePath: getFilePath('archive.zip'),
 			destinationDirectoryPath: getFilePath('destination'),
@@ -139,7 +139,7 @@ describe('Bridge.handleMessage', function () {
 	it('should process CreateDirectory message', async function () {
 		const createDirectory = sinon.stub().resolves();
 		const fileSystem = { createDirectory };
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: CreateDirectory,
 			directoryPath: getFilePath('directory'),
 		});
@@ -151,7 +151,7 @@ describe('Bridge.handleMessage', function () {
 		const fileSystem = {
 			isDirectory: async () => true,
 		};
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: IsDirectory,
 			filePath: getFilePath('directory'),
 		});
@@ -178,7 +178,7 @@ describe('Bridge.handleMessage', function () {
 		const fileSystem = {
 			listStorageUnits: async () => storageUnits,
 		};
-		const result = await handleMessage(fileSystem as any, {} as any, {
+		const result = await handleMessage(fileSystem as any, {} as any, {} as any, {
 			type: ListStorageUnits,
 		});
 		result.should.deepEqual({ storageUnits });