+++ test/unit/Driver/FrontDriverSpec.ts
@@ -7,6 +7,18 @@ import {
 	GetModel,
 	SystemReboot,
 } from '../../../src/Bridge/bridgeSystemMessages';
+import ISocket from '@signageos/front-display/es6/Socket/ISocket';
+
+function createWindow(override: any = {}): Window {
+	return {
+		addEventListener: sinon.spy(),
+		...override,
+	} as any;
+}
+
+function createMockSocket(): ISocket {
+	return new EventEmitter() as any;
+}
 
 describe('Driver.FrontDriver', function () {
 
@@ -20,7 +32,7 @@ describe('Driver.FrontDriver', function () {
 				socketClient: new EventEmitter(),
 			};
 
-			const frontDriver = new FrontDriver({} as any, '1.0.0', bridge as any, 'http://localhost:8081');
+			const frontDriver = new FrontDriver(createWindow(), 'hug', '1.0.0', bridge as any, createMockSocket(), 'http://localhost:8081');
 			const model = await frontDriver.getModel();
 			model.should.equal('model1');
 		});
@@ -34,7 +46,7 @@ describe('Driver.FrontDriver', function () {
 				socketClient: new EventEmitter(),
 			};
 
-			const frontDriver = new FrontDriver({} as any, '1.0.0', bridge as any, 'http://localhost:8081');
+			const frontDriver = new FrontDriver(createWindow(), 'hug', '1.0.0', bridge as any, createMockSocket(), 'http://localhost:8081');
 			await frontDriver.appReboot();
 			bridge.invoke.callCount.should.equal(1);
 			bridge.invoke.getCall(0).args[0].should.deepEqual({ type: SystemReboot });
@@ -44,16 +56,16 @@ describe('Driver.FrontDriver', function () {
 	describe('appRestart', function () {
 
 		it('should invoke restarting of the application', async function () {
-			const window = {
+			const window: any = createWindow({
 				location: {
 					reload: sinon.spy(),
 				},
-			};
+			});
 			const bridge = {
 				socketClient: new EventEmitter(),
 			};
 
-			const frontDriver = new FrontDriver(window as any, '1.0.0', bridge as any, 'http://localhost:8081');
+			const frontDriver = new FrontDriver(window as any, 'hug', '1.0.0', bridge as any, createMockSocket(), 'http://localhost:8081');
 			await frontDriver.appRestart();
 			window.location.reload.callCount.should.equal(1);
 		});
@@ -63,7 +75,7 @@ describe('Driver.FrontDriver', function () {
 
 		it('should invoke restarting of the application', async function () {
 			const bridge = { socketClient: new EventEmitter() };
-			const frontDriver = new FrontDriver({} as any, '1.0.0', bridge as any, 'http://localhost:8081');
+			const frontDriver = new FrontDriver(createWindow(), 'hug', '1.0.0', bridge as any, createMockSocket(), 'http://localhost:8081');
 			const applicationVersion = await frontDriver.getApplicationVersion();
 			applicationVersion.should.equal('1.0.0');
 		});
@@ -73,14 +85,14 @@ describe('Driver.FrontDriver', function () {
 
 		it('should bind callback to key presses', function () {
 			const eventEmitter = new EventEmitter();
-			const window = {
+			const window = createWindow({
 				addEventListener: (type: string, listener: any) => eventEmitter.addListener(type, listener),
-			};
+			});
 			const bridge = {
 				socketClient: new EventEmitter(),
 			};
 
-			const frontDriver = new FrontDriver(window as any, '1.0.0', bridge as any, 'http://localhost:8081');
+			const frontDriver = new FrontDriver(window as any, 'hug', '1.0.0', bridge as any, createMockSocket(), 'http://localhost:8081');
 			const callback = sinon.spy();
 			frontDriver.bindKeyUp(callback);
 			eventEmitter.emit('keyup', { keyCode: 0x25 });